==> scripts/05_geneticAnalysis_01_qtl_01_outcome.R <==
# QTL analysis, using Javi and Nienke - QTL pipeline

setwd('/vol/projects/CIIM/Influenza/ZirrFlu/genetics_abqtl/out/mapping/') # folder with the last updated result

# the locus zoom plot is at : 
# /vol/projects/CIIM/meta_cQTL/QTL_dev/dev/resources/post/locuszoom.R
