==> scripts/05_geneticAnalysis_02_omicspred_03_sumUp_outcome.R <==
rm(list = ls())

library(tidyverse)
library(stringr)

# model information =======================================================================
models <- c("Metabolon", "Nightingale", "Olink", "RNAseq", "Somalogic")
models_info <- list()
for (model_name in models) {
  models_info[[model_name]] <- read.csv2(paste0("data/omicsPred/" , model_name,
                                                "_trait_validation_results_with_OMICSPRED_ID.csv"), sep = "\t")
}

models_Tab_temp <- models_info %>% bind_rows(.id = "model") %>% 
  mutate(type = ifelse(model == "Metabolon"| model == "Nightingale", "metabolite",
                       ifelse(model == "Olink"| model == "Somalogic", "protein", "RNAseq")))
models_Tab_temp %>% count(type)

## select metabolites matched with endogenous metabolites, HMDB database --------------------------
hmdb_endogenous <- read_csv("data/20221011_HMDB_endogenousMetabolites")

models_Tab  <- models_Tab_temp  %>% 
  mutate(Name = ifelse(is.na(Biochemical.Name), Biomarker.Name, Biochemical.Name), # merge the name, pathway/group information of 2 platform together
         subClass = ifelse(is.na(Sub.Pathway), Subgroup, Sub.Pathway),
         class = ifelse(is.na(Super.Pathway), Group, Super.Pathway),
         name = tolower(Name)) %>% 
  dplyr::select(-c("Biomarker.Name", "Biochemical.Name", 
            "Subgroup", "Sub.Pathway", "Group", "Super.Pathway")) %>%
  relocate(c(Metabolon.ID, Trait.ID), .after = name) %>%
  left_join(hmdb_endogenous %>% mutate(name = tolower(NAME))) # annotate to the hmdb database using matched names

models_mebo_formula <- models_Tab  %>% 
  dplyr::select(model, name, CHEMICAL_FORMULA) %>% drop_na() %>% add_count(CHEMICAL_FORMULA)

models_mebo_formula %>% count(model)
models_mebo_formula %>% filter(n > 1)

models_Tab %>% count(model)
rm(hmdb_endogenous, models_info, models_mebo_formula, models_Tab_temp)

# model predicted outcomes (infomation, predicted value, SNP used) ======================================================
models <- c("Metabolon", "Nightingale", "Olink", "RNAseq", "SomaScan") # SomaScan == Somalogic, OmicsPred use different names in model and summary files

predOut <- list()
for (model_name in models) {
  filesList <- list.files(paste0("processedDat/omicsPred/output/", model_name))

  for (fileName in filesList[grep("log", filesList)]) {
    if (file.size(paste0("processedDat/omicsPred/output/", model_name, "/", fileName)) > 0) {
      predOut$log[[substring(fileName, 1, 10)]] <- read.delim(
        paste0("processedDat/omicsPred/output/", model_name, "/", fileName), header = FALSE)}}


  for (fileName in filesList[grep("score", filesList)]) {
    if (grepl("sscore.vars", fileName)) {
      if (file.size(paste0("processedDat/omicsPred/output/", model_name, "/", fileName)) > 0) {
        predOut$var_rsid[[substring(fileName, 1, 10)]] <- read.delim(
          paste0("processedDat/omicsPred/output/", model_name, "/", fileName), header = FALSE)}
      } else {
        if (file.size(paste0("processedDat/omicsPred/output/", model_name, "/", fileName)) > 0) {
          predOut$score[[substring(fileName, 1, 10)]] <- read.delim(
            paste0("processedDat/omicsPred/output/", model_name, "/", fileName))}}
    }
}

# information of the prediction model (.log) ----------------------------------------
# View(predOut$log$OPGS004208) 

pred_logTab <- predOut$log %>% 
  lapply(function(x) # make all logFile size to 25, by adding NA row to files have less than 25 rows
    if(nrow(x) == 24) {x %>% add_row(V1 = NA, .before = 19)
    } else if (nrow(x) == 21) {
      x %>% add_row(V1 = rep(NA, 4), .before = 21)} else if (nrow(x) == 20) {
        x %>% add_row(V1 = rep(NA, 5), .before = 20)} else x) %>%
  imap(~.x %>% rename_with(function(x) .y)) %>% 
  purrr::reduce(cbind)
# pred_logTab[c(1:4, 7:8, 11:18), ] %>% unlist() %>% unique() # return same value across columns, because it is about how the models run
# pred_logTab[24, ] %>% unlist() %>% unique() # return "." (the line from model log file) or NA (filled row in the previous step to match the universe 25 rows on log files)

# Note: extract the variants (SNP) used information from models'log file
pred_logInfo <- pred_logTab[19:20, ] %>% 
  t() %>% as.data.frame() %>% 
  dplyr::rename("warning" = "19", "SNPs_use_info" = "20") %>%
  mutate(num_SNPs_use = ifelse(SNPs_use_info == "Error: No valid variants in --score file.", 0, 
                               str_extract(SNPs_use_info, "[[:digit:]]+")) %>% as.numeric) %>%
  mutate(numSNPs_skip = ifelse(is.na(warning), 0, # NA = skip SNPs is not applicable
                               str_extract(warning, "[[:digit:]]+")) %>% as.numeric) %>% 
  rownames_to_column("OMICSPRED.ID")

# Note: model prediction information 
models_pred_Tab <- models_Tab  %>% 
  dplyr::select(OMICSPRED.ID, X.SNP, model, type, Gene, Ensembl.ID,
         CHEMICAL_FORMULA, name, Name, subClass, class, 
         matches("R2|Rho|MissingRate")) %>%
  full_join(pred_logInfo) %>% relocate(c(warning, matches("SNPs")), .after = "X.SNP")

rm(pred_logInfo, pred_logTab, fileName, filesList, model_name, models)

# select models to further analysis ======================================================
selected_models <- models_pred_Tab %>% filter(num_SNPs_use >= 5) # the R2 can be good?

models_pred_Tab %>% count(type, model)
selected_models %>% count(type, model)

## load predicted value ------------------------------------------------------------------
pred_Tab <- predOut$score %>%
  lapply(function(x) x %>% column_to_rownames("X.IID")) %>% 
  imap(~.x %>% rename_with(function(x) .y)) %>% 
  purrr::reduce(cbind) %>% dplyr::select(selected_models$OMICSPRED.ID)

save(selected_models, pred_Tab, file = "processedDat/omicsPred_output.RData")

# plots for OmicPred models ======================================================
models_count <- models_Tab %>% full_join(models_pred_Tab) %>%
  mutate(status = ifelse(num_SNPs_use >=5, "used", "non_used")) %>%
  mutate(status = ifelse(is.na(status), "used", status)) %>%
  group_by(model, status) %>% summarise(num_models = n())

modelsCount_plot <- models_count %>% 
  ggplot(aes(x = model, y = num_models, fill = status, label = num_models)) + 
  geom_bar(stat = "identity") +
  geom_text(size = 6, position = position_stack(vjust = 0.5)) +
  scale_y_log10(limits = c(1, 1e8), breaks = c(1,1e1, 1e2, 1e3, 1e4, 1e5, 1e6, 1e7, 1e8)) + 
  annotation_logticks(sides = "l")  + 
  scale_fill_manual(values=c( c("#C7C2A6", "#898366"))) +
  theme_classic()+
  theme(text = element_text(size = 18), legend.position="top")

modelsCount_plot


usedModelsCount_plot <- models_count %>% 
  filter(status == "used") %>%
  mutate(model = factor(model, levels = c("RNAseq", "Somalogic", "Olink", 
                                          "Metabolon", "Nightingale"))) %>%
  ggplot(aes(x = model, y = num_models, label = num_models)) + 
  geom_bar(stat = "identity", fill = "#C7C2A6") +
  geom_text(size = 6, vjust=-0.5) +
  scale_y_log10(limits = c(1, 1e4), breaks = c(1,1e1, 1e2, 1e3, 1e4)) + 
  annotation_logticks(sides = "l")  + 
  ylab("Number of molecules predicted by omicsPred") +
  theme_classic()+
  theme(text = element_text(size = 18))

usedModelsCount_plot

# save the plot 
png("output/omicPred_modelsCount_barplot.png", width = 576)
modelsCount_plot
dev.off()


png("output/omicPred_usedModelsCount_barplot.png", height = 432, width = 576)
usedModelsCount_plot
dev.off()

## selected models, plot  / visualization -------------------------------------------------
models_pred_Tab %>% 
 # filter(type == "RNAseq") %>% 
#  filter(type != "RNAseq") %>% 
 # filter(X.SNP < 10) %>%
  ggplot(aes(x = X.SNP, y = as.numeric(Internal_R2))) + 
  geom_point() + theme_classic()

models_pred_Tab %>% 
  filter(type == "RNAseq") %>% 
#  filter(type != "RNAseq") %>% 
  filter(num_SNPs_use < 10) %>%
  ggplot(aes(x = num_SNPs_use, y = as.numeric(Internal_R2))) + 
  geom_point() + theme_classic()

# plot use and non-used models
models_pred_Tab %>% 
  ggplot() + 
  geom_jitter(aes(x = X.SNP, y = as.numeric(Internal_R2), col = model), width = 0.2, size = 1, alpha = 0.5) + 
  scale_x_log10() +
  annotation_logticks(sides = "b")  + 
  geom_vline(xintercept = 5, linetype = "dashed", color = "black") +
  theme_classic()

selectedModels_plot <- models_pred_Tab %>% 
  ggplot(aes(x = X.SNP, y = as.numeric(Internal_R2))) + 
  geom_jitter(aes(color = 
                    factor(ifelse(X.SNP >5, model, "not used"))), 
              width = 0.1, size = 3, alpha = 0.5) + 
  scale_color_manual(values = c("grey", "#B65008", "#751915", "#65771E", "#9B2727", "#036879", "#898366"),
                     breaks = c("not used", "RNAseq", "Olink", "Somalogic", "Metabolon", "Nightigale"),
                     labels = c("not used", "RNAseq", "Olink", "Somalogic", "Metabolon", "Nightigale")) + # Adjust colors as needed
  labs(color = "Model") +
  scale_x_log10() +
  annotation_logticks(sides = "b")  + 
  geom_vline(xintercept = 5, linetype = "dashed", color = "black") +
  theme_classic()+
  theme(text = element_text(size = 18), legend.position="top")

selectedModels_plot

# save the plot 
png("output/omicPred_selectedModels_scatterplot.png", width = 528)
selectedModels_plot
dev.off()
